# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerCatalyst API",
    "description": "Multi-tenant AI-driven SaaS for ATS-optimized resume tailoring",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/resume/analyze-jd": {
      "post": {
        "tags": [
          "Resume Tailoring & RAG"
        ],
        "summary": "Analyze target Job Description",
        "description": "Extract key requirements, technical skills, core responsibilities,\nand ATS keywords from a raw Job Description.",
        "operationId": "analyze_jd_endpoint_api_resume_analyze_jd_post",
        "parameters": [
          {
            "name": "save_jd",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Persist target JD in database",
              "default": false,
              "title": "Save Jd"
            },
            "description": "Persist target JD in database"
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JDAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JDAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/tailor": {
      "post": {
        "tags": [
          "Resume Tailoring & RAG"
        ],
        "summary": "Generate ATS-tailored resume via RAG pipeline",
        "description": "Full end-to-end RAG Resume Tailoring Pipeline delegated to resume_service:\n1. Analyzes the target Job Description.\n2. Retrieves top-K matching candidate achievements via pgvector with multi-tenant isolation.\n3. Fetches candidate base profile.\n4. Synthesizes tailored resume in structured JSON and LaTeX using STAR format.\n5. Calculates objective ATS match score and keyword gap analysis.\n6. Persists the tailored resume record in the database.",
        "operationId": "tailor_resume_endpoint_api_resume_tailor_post",
        "parameters": [
          {
            "name": "save_to_db",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Persist synthesized resume in DB",
              "default": true,
              "title": "Save To Db"
            },
            "description": "Persist synthesized resume in DB"
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TailorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TailoredResumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/history": {
      "get": {
        "tags": [
          "Resume Tailoring & RAG"
        ],
        "summary": "Retrieve candidate generated resumes history",
        "description": "List all previously generated tailored resumes for the tenant and user.",
        "operationId": "get_resume_history_api_resume_history_get",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TailoredResumeModelResponse"
                  },
                  "title": "Response Get Resume History Api Resume History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/{resume_id}": {
      "get": {
        "tags": [
          "Resume Tailoring & RAG"
        ],
        "summary": "Retrieve a specific tailored resume",
        "description": "Fetch a specific tailored resume by ID with tenant verification.",
        "operationId": "get_resume_by_id_api_resume__resume_id__get",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TailoredResumeModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JDAnalysis": {
        "properties": {
          "job_title": {
            "type": "string",
            "title": "Job Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "primary_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Skills",
            "default": []
          },
          "core_responsibilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Core Responsibilities",
            "default": []
          },
          "keywords_to_target": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords To Target",
            "default": []
          },
          "seniority_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority Level"
          }
        },
        "type": "object",
        "required": [
          "job_title"
        ],
        "title": "JDAnalysis",
        "description": "Structured extraction of target job description requirements via Gemini."
      },
      "JDAnalysisRequest": {
        "properties": {
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          }
        },
        "type": "object",
        "required": [
          "job_description"
        ],
        "title": "JDAnalysisRequest",
        "description": "Payload to request automated analysis of a job description."
      },
...
      "SynthesizedExperience": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "tailored_bullets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tailored Bullets"
          }
        },
        "type": "object",
        "required": [
          "company",
          "role",
          "tailored_bullets"
        ],
        "title": "SynthesizedExperience",
        "description": "Experience item tailored specifically to target JD keywords."
      },
      "TailorRequest": {
        "properties": {
          "job_description": {
            "type": "string",
            "minLength": 10,
            "title": "Job Description",
            "description": "Full job description text"
          },
          "target_job_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Job Title",
            "description": "Target job title override"
          },
          "top_k_bullets": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K Bullets",
            "description": "Top-K bullets to retrieve via pgvector RAG"
          },
          "master_bullets": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Master Bullets",
            "description": "Direct bullets override bypassing RAG retrieval"
          }
        },
        "type": "object",
        "required": [
          "job_description"
        ],
        "title": "TailorRequest",
        "description": "Request payload to initiate AI RAG tailoring pipeline."
      },
      "TailoredResumeContent": {
        "properties": {
          "candidate_name": {
            "type": "string",
            "title": "Candidate Name"
          },
          "contact_info": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Contact Info"
          },
          "professional_summary": {
            "type": "string",
            "title": "Professional Summary"
          },
          "highlighted_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Highlighted Skills"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/SynthesizedExperience"
            },
            "type": "array",
            "title": "Experiences"
          },
          "projects": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Projects",
            "default": []
          },
          "education": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Education",
            "default": []
          },
          "certifications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Certifications",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "candidate_name",
          "contact_info",
          "professional_summary",
          "highlighted_skills",
          "experiences"
        ],
        "title": "TailoredResumeContent",
        "description": "Full structured JSON representation of a synthesized tailored resume."
      },
      "TailoredResumeModelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "jd_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jd Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "structured_content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Structured Content"
          },
          "raw_latex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Latex"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "title",
          "structured_content",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "TailoredResumeModelResponse",
        "description": "Database record response for a saved tailored resume."
      },
      "TailoredResumeResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "job_title": {
            "type": "string",
            "title": "Job Title"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "matched_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Keywords",
            "default": []
          },
          "missing_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Keywords",
            "default": []
          },
          "structured_resume": {
            "$ref": "#/components/schemas/TailoredResumeContent"
          },
          "latex_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latex Source"
          }
        },
        "type": "object",
        "required": [
          "job_title",
          "structured_resume"
        ],
        "title": "TailoredResumeResponse",
        "description": "Final synthesized resume response with ATS match score and LaTeX."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}